added dircheck to rex_dir::delete() to get around php warnings resulting from an DirectoryIterator of a not existant folder.

+++ redaxo/src/5.0/core/lib/util/dir.php
@@ -66,7 +66,7 @@ class rex_dir
    */
   static public function delete($dir, $deleteSelf = true)
   {
-    if(self::deleteIterator(self::recursiveIterator($dir)))
+    if(is_dir($dir) && self::deleteIterator(self::recursiveIterator($dir)))
     {
       if(!$deleteSelf || rmdir($dir))
       {
